+++ src/parakeet/analyse/_extract.py
@@ -19,6 +19,10 @@ from typing import Any
 from functools import singledispatch
 from math import sqrt, ceil
 
+
+__all__ = ["extract"]
+
+
 # Set the random seed
 random.seed(0)
