+++ src/parakeet/sample/__init__.py
@@ -1986,9 +1986,8 @@ def load(filename: str, mode: str = "r") -> Sample:
 
 
 # fmt: off
-from parakeet.sample.new import new # noqa
-from parakeet.sample.add_molecules import add_molecules # noqa
-from parakeet.sample.add_molecules import add_single_molecule # noqa
-from parakeet.sample.mill import mill # noqa
-from parakeet.sample.sputter import sputter # noqa
+from parakeet.sample._new import * # noqa
+from parakeet.sample._add_molecules import * # noqa
+from parakeet.sample._mill import * # noqa
+from parakeet.sample._sputter import * # noqa
 # fmt: on
