+++ src/parakeet/simulate/__init__.py
@@ -1,8 +1,8 @@
 # fmt: off
-from parakeet.simulate.projected_potential import * # noqa
-from parakeet.simulate.exit_wave import * # noqa
-from parakeet.simulate.optics import * # noqa
-from parakeet.simulate.image import * # noqa
-from parakeet.simulate.ctf import * # noqa
-from parakeet.simulate.simple import * # noqa
+from parakeet.simulate._projected_potential import * # noqa
+from parakeet.simulate._exit_wave import * # noqa
+from parakeet.simulate._optics import * # noqa
+from parakeet.simulate._image import * # noqa
+from parakeet.simulate._ctf import * # noqa
+from parakeet.simulate._simple import * # noqa
 # fmt: on
