+++ src/parakeet/simulate/_exit_wave.py
