+++ src/parakeet/simulate/_image.py
