+++ src/parakeet/simulate/_optics.py
