+++ src/parakeet/simulate/_projected_potential.py
